eCommerce Product Catalog 2.6.5
eCommerce Product Catalog 2.6.5

+++ modules/attributes/product-attributes.php
@@ -19,7 +19,6 @@ add_action( 'init', 'ic_create_product_attributes' );
  *
  */
 function ic_create_product_attributes() {
-
 	$args		 = array(
 		'label'			 => 'Attributes',
 		'hierarchical'	 => true,
@@ -27,7 +26,7 @@ function ic_create_product_attributes() {
 		'query_var'		 => false,
 		'rewrite'		 => false,
 	);
-	$post_types	 = apply_filters( 'ic_attributes_register_post_types', array( 'al_product' ) );
+	$post_types	 = apply_filters( 'ic_attributes_register_post_types', product_post_type_array() );
 	register_taxonomy( 'al_product-attributes', $post_types, $args );
 }
 
@@ -78,6 +77,7 @@ function ic_assign_product_attributes( $product_meta, $post, $clear_empty = true
 	if ( $max_attr > 0 ) {
 		$product_id	 = isset( $post->ID ) ? $post->ID : $post;
 		$attr_ids	 = array();
+
 		for ( $i = 1; $i <= $max_attr; $i++ ) {
 			if ( empty( $product_meta[ '_attribute' . $i ] ) || (isset( $product_meta[ '_attribute' . $i ][ 0 ] ) && empty( $product_meta[ '_attribute' . $i ][ 0 ] )) ) {
 				continue;
@@ -87,10 +87,12 @@ function ic_assign_product_attributes( $product_meta, $post, $clear_empty = true
 				if ( !empty( $label ) ) {
 					$value = is_array( $product_meta[ '_attribute' . $i ] ) ? ic_sanitize_product_attribute( $product_meta[ '_attribute' . $i ][ 0 ] ) : ic_sanitize_product_attribute( $product_meta[ '_attribute' . $i ] );
 					if ( !empty( $value ) ) {
-						$label_id	 = ic_add_product_attribute_label( $label );
-						$attr_ids[]	 = $label_id;
-						$value_id	 = ic_add_product_attribute_value( $label_id, $value );
-						$attr_ids[]	 = $value_id;
+						$label_id = ic_add_product_attribute_label( $label );
+						if ( !empty( $label_id ) ) {
+							$attr_ids[]	 = $label_id;
+							$value_id	 = ic_add_product_attribute_value( $label_id, $value );
+							$attr_ids[]	 = $value_id;
+						}
 					}
 				}
 			}
@@ -106,6 +108,8 @@ function ic_assign_product_attributes( $product_meta, $post, $clear_empty = true
 	return $product_meta;
 }
 
+add_action( 'ic_scheduled_attributes_clear', 'ic_clear_empty_attributes' );
+
 /**
  * Clears empty product attributes
  *
@@ -117,12 +121,24 @@ function ic_clear_empty_attributes() {
 		'hide_empty' => 0,
 		'number'	 => $max_attr
 	) );
-	foreach ( $attributes as $attribute ) {
-		if ( $attribute->count == 0 ) {
-			wp_delete_term( $attribute->term_id, 'al_product-attributes' );
-		} else {
-			break;
+	$schedule	 = false;
+	if ( !empty( $attributes ) && is_array( $attributes ) && !is_wp_error( $attributes ) ) {
+		foreach ( $attributes as $attribute ) {
+			if ( $attribute->count == 0 ) {
+				$schedule = true;
+				wp_delete_term( $attribute->term_id, 'al_product-attributes' );
+			} else {
+				$schedule = false;
+				break;
+			}
 		}
+		if ( !wp_get_schedule( 'ic_scheduled_attributes_clear' ) && $schedule ) {
+			wp_schedule_event( time(), 'hourly', 'ic_scheduled_attributes_clear' );
+		} else {
+			wp_clear_scheduled_hook( 'ic_scheduled_attributes_clear' );
+		}
+	} else {
+		wp_clear_scheduled_hook( 'ic_scheduled_attributes_clear' );
 	}
 }
 
@@ -194,6 +210,17 @@ function ic_system_tools_attributes_upgrade() {
 		}
 		echo '</td></tr>';
 	}
+	if ( wp_get_schedule( 'ic_scheduled_attributes_clear' ) ) {
+		if ( isset( $_GET[ 'clear_products_attributes' ] ) ) {
+			ic_clear_empty_attributes();
+		}
+	}
+	if ( wp_get_schedule( 'ic_scheduled_attributes_clear' ) ) {
+		echo '<tr>';
+		echo '<td>Clear Attributes</td>';
+		echo '<td><a class="button" href="' . admin_url( 'edit.php?post_type=al_product&page=system.php&clear_products_attributes=1' ) . '">Speed UP Clearing Empty Attributes</a></td>';
+		echo '</tr>';
+	}
 }
 
 /**
@@ -217,7 +244,7 @@ function ic_get_attribute_id( $label ) {
  */
 function ic_get_attribute_name( $attribute_id ) {
 	$attribute = get_term_by( 'id', $attribute_id, 'al_product-attributes' );
-	if ( $attribute ) {
+	if ( $attribute && $attribute->count > 0 ) {
 		return $attribute->name;
 	}
 	return false;
@@ -229,18 +256,31 @@ function ic_get_attribute_name( $attribute_id ) {
  * @param type $label
  * @return boolean
  */
-function ic_get_attribute_values( $label ) {
+function ic_get_attribute_values( $label, $format = 'names' ) {
 	$attribute_id	 = ic_get_attribute_id( $label );
-	$values			 = get_term_children( $attribute_id, 'al_product-attributes' );
+	//$values			 = get_term_children( $attribute_id, 'al_product-attributes' );
+	$values			 = get_terms( array(
+		'taxonomy'	 => 'al_product-attributes',
+		'hide_empty' => true,
+		'parent'	 => $attribute_id,
+		'fields'	 => $format
+	) );
 	if ( empty( $values ) || is_wp_error( $values ) || !is_array( $values ) ) {
 		return false;
 	}
-	$attributes = array();
-	foreach ( $values as $value_id ) {
-		if ( !empty( $value_id ) ) {
-			$attributes[] = ic_get_attribute_name( $value_id );
-		}
-	}
+	/*
+	  $attributes = array();
+	  foreach ( $values as $value_id ) {
+	  if ( !empty( $value_id ) ) {
+	  $attr_name = ic_get_attribute_name( $value_id );
+	  if ( $attr_name ) {
+	  $attributes[] = $attr_name;
+	  }
+	  }
+	  }
+	 *
+	 */
+	$attributes = $values;
 	return $attributes;
 }
 
@@ -251,7 +291,7 @@ function ic_get_attribute_values( $label ) {
  * @return type
  */
 function ic_sanitize_product_attribute( $attribute ) {
-	$sanitized_attribute = wp_unslash( sanitize_term_field( 'name', $attribute, 0, 'al_product-attributes', 'db' ) );
+	$sanitized_attribute = trim( wp_unslash( sanitize_term_field( 'name', $attribute, 0, 'al_product-attributes', 'db' ) ) );
 	if ( strlen( $sanitized_attribute ) > 200 ) {
 		return '';
 	}
@@ -279,6 +319,6 @@ function ic_delete_all_attribute_terms() {
  * @return type
  */
 function get_all_attribute_labels() {
-	$attributes_labels = get_terms( 'al_product-attributes', array( 'parent' => 0, 'fields' => 'names' ) );
+	$attributes_labels = get_terms( 'al_product-attributes', array( 'parent' => 0, 'fields' => 'names', 'hide_empty' => true ) );
 	return $attributes_labels;
 }
